This is a bugfixed installer for Minecraft.
This bugfixed installer for Minecraft, the .deb installer, resolves the error "Dependency is not satisfiable: libgdk-pixbuf2.0-0. This error is due to a missing hyphen between "pixbuf" and "2." I repackaged the .deb file with the proper dependency name, tested it on hardware, and it performed without any errors. I hope it serve you well, particularly new and longtime Debian users.